fix(mount): Fix readahead under high latency (#514)
This change targets fixing some crashes of the clients while using the
readahead algorithm on some slow-to-respond clusters. These crashes
were caused in older versions due to accessing invalid cache entries
and in the current state of the code due to reuse of some cache entries
that make its state inconsistent.

The main case where this happened was the following:
- some ReadRecord scheduled several read requests in advance.
- the first read requests to be satisfied is not the one with least
offset (and first in the order of schedule), so it needs to wait until
all the previous ones get satisfied.
- that first resolved request waits too much and gets expired, and the
released (or sent to the entries pool), but not taken away from the
ReadRecords pending requests container.
- it is then added to some read reply (from the pending requests
container) while being invalid.

The proposed solution is to remove this kind of entries from the cache
(such that they won't appear for queries) but keeping them alive for as
long as they are not removed from the pending requests container.

Some consistency checks were also added.

Merge pull request #514 from leil-io/fix-high-latency-reads

Co-authored-by: Rolando Sánchez Ramos <rolysr@leil.io>
Signed-off-by: Dave <dave@leil.io>