fix(cs): Fix release of failed read jobs
This change was motivated by a high memory consumption of the
chunkserver processes in some scenarios where some failed reads are
involved.

The cause of the issue is that entries with failed reads that were
already set to be deleted, stayed alive because they were waiting for
discarded jobs callbacks that were already processed (right after the
fail).

The changes can be split in the following:
- set the OutputBuffer status of the reads failing in the `hddOpen`
(see `job_read`). This allows all failing read jobs to have the failed
status set.
- while discarding pending read jobs while handling a failed read, the
jobs with statuses set should not be moved to the new container: these
jobs should have been already handled and can be already destryed and
move the structures to the BuffersPool.

As a side change, to have a better look at how it is looking the
chunkserver memory we will start tracking the number of blocks in
iobuffers and log them (log_debug level).

Signed-off-by: Dave <dave@leil.io>