feat(meta): Optimize readdir for meta mounts
The previous implementation of the metadata mountpoint readdir
operation was inefficient, as it retrieved all trash and reserved
files from the master in a single protocol call. This could lead
to excessive memory usage and slow execution times. This commit
improves the process by retrieving only predefined limited amount
of entries data from the client side, making it more flexible to
integrate other trash/reserved files improvement features on top
of this one.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>