fix(cs): Fix SYSLOG_IDENT reload
Previously, after performing a reload on a chunkserver, the
SYSLOG_IDENT (i.e sfschunkserverX for instance) of its logs got lost.
Also, the thread id of the producer of the log, was not getting logged
as well. This commit fixes it by performing the setup of those
parameters after the logs are setup during the reload: the setup of the
logs destroyed and created the underlying logs structure.

Signed-off-by: Dave <dave@leil.io>