fix(cs): Keep alive disks with pending chunks
This commit fixes a possible crash of the chunkserver if being reloaded
to remove a drive from its hdd config file. The issue can be described
as the following:
- some disk is removed from the hdd config file and the CS is reloaded
- the master is updated with the list of new missing chunk parts and
the chunks are being removed from the CS structures unless it is under
use.
- in any case, the disk structure is released.
- the chunks in use in most of the cases assume that its owner (the
recently removed disk) is valid and this causes segfault error and the
CS crash.

The fix consists of moving the disks instances removed from config to
some other container, different from the usual gDisks, and wait for the
pending chunks in use to release those structures.

Signed-off-by: Dave <dave@leil.io>