fix(mount): Fix mount info on master reconnection (#527)
The previous master-side mount-info list request could fail to return
expected values after a master restart or reconnection, which can
occur after a master upgrade in an HA cluster. This commit fixes the
issue by ensuring that saunafs-admin and the CGI monitoring interface
receive up-to-date mount information for each client session as soon
as the master comes back online.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>