fix(cs): Fix release of pending reads
It has being noticed that in some read workloads the memory used by the
chunkserver can increase steadily over time. The issue is related to
the proper release of the structures related to the reads in the
chunkserver side. This change targets fixing that issue and some other
inconsistencies found while studying the issue.

The main change that fixed the issue was removing the actually disabled
jobs in the jobs pool from the ones put in the discarded read jobs
container. Those jobs were not setting up the status of the related
OutputBuffer and thus, were getting stuck the releasing mechanism of
that structure and the whole csentry.

There were a couple of fixes applied along the way:
- in the function changeCallback, while called upon a list of jobs, the
std::move of the callback provided was substituted by a direct copy.
- in the function readDiscardCallback make sure there are no jobs with
status after the cycle.

Beside these fixes, the other important change is to simplify a bit the
closing scheme of the csentry.

Signed-off-by: Dave <dave@leil.io>