fix(chunkserver): Don't clear chunkserver output packets
When the master connection was closed (for example, when it's demoting),
it could cause inconsistent state with the master. If a chunkserver
received a packet, to create a chunk for example, it will finish it but
not reply to master. This could lead to lost data if there were active
writes happening at the same time.

This fix ensures that wehn chunkserver reconnects with master, the
chunkserver will always send the register packet first, and then send
the remaining output packets still remained in the output packets when
the disconnect happened.

Signed-off-by: Urmas Rist <urmas@leil.io>