feat(meta): Add limitless meta mounts readdirs
Previously, readdir operations on the meta mountpoints were limited by
a 100K entries cap on the client side for both trash and reserved file
folders. Additionally, those 100K returned entries were always the
first ones in the corresponding containers on the master side that
store trash and reserved files data.

This commit adds new auxiliary containers on the master side, which
maintain an ordered version of the trash and reserved entries by
uniquely generated IDs for each of them. This allows the meta
mountpoints to consistently request batches of 100K entries from the
master on demand, enabling the listing of all available trash or
reserved file entries and avoiding overstressing the master process
and memory consumption peaks when responding to these requests.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>