chore: Make master version global on client
When it was needed to add new specific client to master
communication or execution logic, it was only possible to
check master version and change logic accordingly on the
mastercomm.cc file. This commit now allows to also trigger
master version specific code in other parts of the client,
as connected master version is not accessible to other parts.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>