fix: Fix writes during crashes/stops (#550)
It was noticed during some upgrade tests that chunks were lost after
restarting the master while writing to the cluster. This issue is
related to the scarse failures of the test
test_network_partition_during_writes. Both of those issues should be
fixed with the proposed changes.

The troublesome situation is the following:
- during the creation of a chunk, the master needs to ask the
chunkservers to create the chunk parts and waits for it.
- when the replies arrive, if succesful, the master replies with OK
status to the client, but if that is not the case then it replies
SAUNAFS_ERROR_NOTDONE.
- this makes the write operation in the client to fail and leaves a
chunk lost behind.

The proposed solution changes the master behavior in the case of a
failed chunk creation:
- the function that replies to the client also removes the chunk from
the file list of chunks.
- the returned status was changed to SAUNAFS_ERROR_CHUNKLOST.

The chunk structure deletion in master will come in the next chunks
loop (it does not have any files pointing to it) and in chunkserver
side right after it.

Signed-off-by: Dave <dave@leil.io>
Co-authored-by: Urmas Rist <urmas@leil.io>