feat: Upgrade SaunaFS FSAL to Ganesha v7.1
This commit upgrades the SaunaFS FSAL to the version released by the
Ganesha community (commit 683de56). The new codebase is compatible
with Ganesha v7.1 and ntirpc v7.1.

In addition to compatibility updates, it brings important fixes and
improvements from upstream, providing a more stable and up-to-date
Ganesha version for SaunaFS.

Signed-off-by: Crash <crash@leil.io>