ci: Use ntirpc from Ganesha submodule
Previously, the ntirpc library was built separately, requiring an extra
build step in the pipeline and additional maintenance for each upgrade.

This commit switches to using the `libntirpc` submodule provided in the
Ganesha repository. As a result, only Ganesha needs to be upgraded, and
ntirpc will automatically be updated through its submodule reference,
reducing complexity and build time.

Signed-off-by: Crash <crash@leil.io>