fix(vcpkg): Make baseline update optional (#571)
The vcpkg_install.sh script is intended to pin the package versions for
achieving reproducible environments. It also sets the version baseline
for packaging. However, updating the baseline each time could
potentially break builds.

This commit allows safer execution of the script by making baseline
updates optional and disabled by default.

It also updates the `vcpkg.json` to include an indirect dependency of
a previous change that manually included a package. This maintains
the list of pinned dependency versions complete. 

Signed-off-by: Baldor <baldor@leil.io>