feat: Read prefetch framework (#564)
This commit introduces the framework for inode-level read prefetching.
It does not yet support any kind of configuration or tuning, but it can
be designed afterwards.

The idea is to create and maintain a new ReadRecords map that is used to
track those readaheaded inode, the gInactiveReadRecords. When a read
request is made, if the inode is found in this map, it is moved to the
active map, gActiveReadRecords, and readahead is performed as usual, but
previously prefetched data should be there already.

It was also added a inodesWithFailedReads structure to track inodes that
had read failures, so that prefetching is not attempted for them. These
structures are being progressively cleaned up by the delayed_ops thread.

Signed-off-by: Dave <dave@leil.io>