perf: Improve connecting files and chunks
This commit replaces the chunk_find() function that always starts
searching at the beginnig of the chunks bucket by a new one starting
from the last visited chunk in the bucket.

This approach is expected to be more efficient, specifically during
the initial loading, where all the chunks are loaded sequentially
for each bucket.

Co-authored-by: Dave <dave@leil.io>

Signed-off-by: Crash <crash@leil.io>