test: Add support for FDB backends
This commit adds initial support for starting the new sfsmds executable
instead of the sfsmaster. sfsmds uses FoundationDB as metadata backend,
so METADATA_BACKEND=FDB must be defined at the beginning of the test.

The testing framework will create an on-demand FDB database and the
initial configuration for the sfsmds.

Co-authored-by: Crash <crash@leil.io>

Signed-off-by: guillex <guillex@leil.io>