feat(master): Add initial id generator interface
This commit starts moving the id generation towards a more extensible
interface. IIdGenerator provides the interface, compatible with the
existing code, while IdGeneratorWithDetainer maintains the current
behavior based on an inode pool detainer.

It is expected some custom implementations to arrive soon to provide a
distributed id generation for the MDS.

Signed-off-by: guillex <guillex@leil.io>