fix: Fix `removeReservedEntry` function usage
Since commit 4386f2dd, the usage of the `removeReservedEntry`
function in the master code was unsafe when called from
`fs_do_emptyreserved`. The execution path could lead to
`removeReservedEntry` being called with a null `FSNodeFile`
instance, resulting in a dereference error when accessing
its `id` field. This commit fixes the issue by improving
`removeReservedEntry` to accept only the required parameters,
avoiding null reference passing on id value. The problem was
detected when compiling the project using the `release-vcpkg`
preset.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>