fix: Fix issue with assert version check
After commit 69eecb26, an assert check in the
`ChunkWorker::tryReplication` function caused an unused
variable warning when building with the `release-vcpkg`
preset. This happened because the variable was only
used in an assertion, which is omitted in release
builds, leading to a compilation error in release-ready
SaunaFS binaries. This commit resolves the issue.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>