refactor: Rework csentry write scheme
This commit targets improving the readability of ChunkserverEntry code
and simplify the write scheme while keeping its main traits.

The main changes readability-wise are the following:
- substitute direct checks for functions having meaningful names.
- substitute duplicated code for new functions.
- modernize all logging code.

The simplification of the write scheme strives in getting rid of the
locking/unlocking/notification/wait extra code related to the fact that
some enqueued InputBuffer could be updated to add some other block of
data. The new approach does not allow enqueued InputBuffers to be
modified: it achieves the multi-block writes by allowing more than one
block to be read into the inputPacket's inputBuffer and several
inputBuffers to be ready for enqueuing at some point with the
writeDataBuffers.

Signed-off-by: Dave <dave@leil.io>