refactor(cs): Rework write operations (#607)
This PR targets improving the readability of ChunkserverEntry code
and simplify the write scheme while keeping its main traits.

The main changes readability-wise are the following:
- substitute direct checks for functions having meaningful names.
- substitute duplicated code for new functions.
- modernize all logging code.
- reduce the packet verbosity.

The simplification of the write scheme strives in getting rid of the
locking/unlocking/notification/wait extra code related to the fact that
some enqueued InputBuffer could be updated to add some other block of
data. The new approach does not allow enqueued InputBuffers to be
modified: it achieves the multi-block writes by allowing more than one
block to be read into the inputPacket's inputBuffer and several
inputBuffers to be ready for enqueuing at some point with the
writeDataBuffers.

Another change in the client side target reducing the number of
lock/unlocks of the inodedata's mutex when writing (buffering) on it,
at the same time the gMutex is not waited on. The main idea is to
protect the truncateLocatorsData using a specific mutex (cannot be
made atomic) instead of the gMutex.

Merge pull request #607 from leil-io/refactor-write-operations

Signed-off-by: Dave <dave@leil.io>