refactor(cs): Remove legacy serializer code
This change targets simplifying the ChunkserverEntry code by removing
all the code considering LegacyMessageSerializer as a possible
MessageSerializer. Taking into consideration that only a
SaunaFsMessageSerializer is the available serializer, thus calls to
static functions can be used.

We are also applying the function createAttachedWriteStatus in a couple
of places where it could fit.

Signed-off-by: Dave <dave@leil.io>