fix: Improve safety for same PID path by inode
This commit improves the path by inode feature safety when for
many path by inode requests, there is the same PID that
triggered it. This scenario could lead to release an entry
reference on the gInodePathInfo structure, when one request
already finished by another one was about to start reading
from stored path for that corresponding entry on that global
container.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>