fix(cs): Fix chunkserver while timing out reads
This commit fixes a couple of concerning issues appearing in
chunkserver while handling a high load of read requests that caused
some of them to time out:
- possible crash of the chunkserver: the already disabled jobs in
prepareDiscardReadJobs were having assigned the readDiscardCallback as
callback, which in cases of fast destruction of the ChunkserverEntry
instance may have caused those callbacks to try using unassigned memory
and therefore SEGFAULT. The fix consists in changing those disabled
jobs callback into not doing anything.
- extra closing chunks ("refcount = 0" messages): this surplus of
job_close relates to assuming some job_read previously scheduled as
successful when they weren't or they were not processed at all. The fix
consists in setting the isChunkOpen member as true when some of those
job_read ends successfully and closing only a single time at the end
if the chunk is said to be open.

Side changes:
- improved the "refcount = 0 - This should never happen!" to include
the troubling chunk.
- assert that ChunkserverEntry state in delayedCloseCallback and
delayedDiscardCallback is State::CloseWait.

Signed-off-by: Dave <dave@leil.io>