feat: Add changelog notifier integration
This commit introduces changes on the master side to allow
registration of a new type of application that receives custom
changelog messages for metadata operations. The implementation
enables external applications to subscribe to metadata events
and receive messages with corresponding file paths resolved by
inode, improving integration and observability for metadata
changes.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>