perf(cs): Streamline blocker operations
This change takes advantage of the new capabilities of the
ProducerConsumerQueue to prioritize the open and get_blocks operations
over the remaining operation types. Those operations needs to be
replied in order to continue sending data in the case of writes (open
case) and to start requesting data for replications (get_blocks case).
This is why those are kind of blocker operations and needs to be
processed faster.

Signed-off-by: Dave <dave@leil.io>