refactor(master): Give own files to matoclserv sessions (#623)
File matoclserv.cc has more than 6 thousands lines, making it difficult
to maintain. This commit moves the code related to session management
to the new files matoclserv_sessions.h and matoclserv_sessions.cc, in
order to segregate better the functionalities related to sessions.

The commit only moves the content to the new files, but avoids any deep
refactor, to ease the review process. A deep refactor is expected soon,
but it is not in the scope of this change.

Signed-off-by: guillex <guillex@leil.io>