fix(cs): Fix replica writes - DISCONNECTED error (#626)
It was noticed that in some scenarios the replica writes might fail,
and the error code sent from the chunkservers is DISCONNECTED. This
is caused due to a misshandled case when writing to a socket from the
InputBuffer code. This function was returning 0: causing an immediate
stop in the forward operation and sending through the chain the
previously mentioned error in some cases when it was supposed to return
-1 and use the errno (EAGAIN) to retry the write. The fix consists in
analizing whether the block write had some previous header data written
to the socket and return the expected error when a block write fails
with EAGAIN errno.

A similar fix was applied preemptively to the reads to the InputBuffer.
Also the missing places where the fwdError function was called were
logged for better future debugging.

Signed-off-by: Dave <dave@leil.io>