feat(chunkserver): Add PLUGINS_DIR config variable
This commit allows to specify a directory (absolute path) to
search for available plugins. For instance:

PLUGINS_DIR = /opt/saunafs/plugins/chunkserver

The Plugin Manager will search in this order:
1. PLUGINS_DIR: defined by the user, takes higher precedence.
2. /usr/local/lib/saunafs/plugins/chunkserver
3. PLUGINS_PATH/chunkserver: defined at configuration time (cmake),
   depends on the build settings (Debug, Release, installDir,
   CMAKE_INSTALL_PREFIX).
4. /usr/lib/saunafs/plugins/chunkserver
5. BUILD_PATH: Useful for development, points to the build directory.

The scan stops after finding plugins in one of the directories, in
order to not mix plugins from different directories.

Co-authored-by: Urmas Rist <urmas@leil.io>

Signed-off-by: guillex <guillex@leil.io>