perf(cs): Configure buffers pool size
This change targets improving the tuning of the chunkserver by setting
the max size of the buffers pool. It allows to reduce the memory used
by the chunkserver if performance is the desired one and to have some
extra buffering space if the memory is not an issue.

The commit also adds a new test to check the new parameter does not
affect the chunkserver behavior.

Signed-off-by: Dave <dave@leil.io>