fix(tools): Restore console code pages on Windows (#640)
Changes from commit 1ffe9bd8 were making the changes made to the console
code pages permanent on the current PowerShell session on Windows where
the command was executed. This could lead to other applications
misbehaving due to unexpected code page settings if executed on the same
console. This commit ensures that the original code pages are restored
after the necessary operations from saunafs command are completed.

Signed-off-by: rolysr <rolysr@leil.io>