refactor(master): make FS quota functions extensible
The remaining quota-related function declarations in filesystem.h were
previously implemented in filesystem_quota.cc, providing only a
concrete implementation assuming that all the metadata related to
quotas was in memory.

This change moves the function declarations to the
IFilesystemOperations interface and its base implementation, to provide
extensibility for the quota-related functions. The current base
implementation (for in-memory metadata) just uses the existing
functions from the newly introduced namespace `quotas`.

Signed-off-by: guillex <guillex@leil.io>