test: Stabilize metadata notifier test (#654)
Previous version of test for metadata notifier behavior was flaky due to
outputing the content generated from the given notifier binary to log
files located on the mountpoint, which could potencially create extra
uneeded noise when expected the operations related with folder access as
checked on the test. This commit fixes that potencial issue.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>