refactor(master): Modernize master-metalogger comm
Previous version of the master to metalogger service communication was
using an old C-styled code, which made this implementation hard to
maintain and understand. Also, previous manual memory assigment and free
was performed manually, which is an error-prone approach. This commit
modernizes the master to metaloggers service by moving to a C++-style
based code.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>