refactor(cs): Encapsulate high level operations
This change targets encapsulating the high level operations logic and
extract it from within the ChunkserverEntry struct. This is done to
pave the way for more complex changes in the chunkserver operations,
such as write buffering.

The high level operation classes aim to expose a relativaly simple
interface and keep for itself most of the logic of each of those
operations. The callbacks will be directly called upon the HLO
instances instead of some ChunkserverEntry passed via parameters. The
new classes need to access certain members of the ChunkserverEntry
struct and this is restrained in the HighLevelOp class to avoid
missuse.

As a side change, we've remove the `extra` parameter from the client
bgjobs functions. It is not necessary anymore.

Signed-off-by: Dave <dave@leil.io>