ci: Run unit tests in Jenkins instead of GH action (#678)
Due to recent issues on the Github action, where FDB tests were failing,
it was decided to move the unit testing back to Jenkins. For a start, a
VM was created specifically to run the unittests. Later on improvements
could be made, for example running in Docker or sharing the build from
the docker image.

The unittests script was also improved to save the XML files, which can
be parsed by junit

Signed-off-by: Urmas Rist <urmas@leil.io>