feat(notifier): Add trash path by inode support (#674)
Previous version of the supported endpoint on the master side for
answering path and type values by a given inode from metadata-notifier
applications was not returning correct path when the requested inode
was on trash or reserved files entries. This commit ensures to add
support for that kind of requests with the expected response and also
improves some variables naming on the
`std::string FilesystemOperationsBase::fullPathByInode` function to be
camel cased.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>