fix(cs): Fix test force read timeouts flakiness (#682)
This change targets fixing the instability of the test
test_chunkserver_force_timeout_reads. The test failed due to sudden
stop of one or several of the chunkservers, causing failed checks in
the end or timeout.

The chunkserver stop seems to be related to inconsistencies in the
assumptions of the jobs to be discarded. Those jobs needs to be not
processed at the moment of adding them to the discarded lists in order
to make pendingDelayedJobs_ hold the correct value. The proposed
solution substitutes the status member of the OutputBuffer for a
isCallbackStarted flag to better maintain whether the callback related
to the OutputBuffer has already started or not.

Side changes:
- the ReadHighLevelOp callbacks will use the second parameter to update
the related OutputBuffer.
- since there will be no status updates from the pool worker threads,
there is no need for a mutex member in OutputBuffer.

Signed-off-by: Dave <dave@leil.io>