feat: Upgrade SaunaFS FSAL to Ganesha v9.2 (#685)
This commit upgrades the SaunaFS FSAL to the version released by the
Ganesha community (commit e353690). The new codebase is compatible
with Ganesha v9.2 and ntirpc v7.2.

In addition to compatibility updates, it brings important fixes and
improvements from upstream, providing a more stable and up-to-date
Ganesha version for SaunaFS.

Signed-off-by: Crash <crash@leil.io>