fix(master): Remove unused case insensitive check (#691)
Adding lowercase entries for each inode and its children during edge
parsing is unnecessary when loading metadata as part of filesystem
initialization on the master. At that point, there is no context
about case-insensitive client connections, so the  case-insensitive
check for each inode and its children always  evaluates to false and
the code never executes. This commit removes that dead code.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>