Fix: Fix overwriting failing tests
Move the next tests to LongSystemTests suite:
-test_xor_overwriting_faulty_chunkservers
-test_xor_overwriting_faulty_master_and_chunkservers
-test_xor_overwriting_faulty_master

Fix the logic in SaunaClient::readdir regarding
updateNextReaddirEntryIndexIfMasterRestarted. After a master
restart updateNextReaddirEntryIndexIfMasterRestarted
function will be called only once if readdirSession->restarted
value doesn't change. Also if this function fails there is no
error handling for it.

Fix bug in the logic of updateNextReaddirEntryIndexIfMasterRestarted,
the nextEntryIndex is moved to the end of the directory when
readdirSession.lastReadIno == 0,the expected behavior is to stay
in the beggining nextEntryIndex=0.

Fix bug on getxattr not using RETRY_ON_ERROR_WITH_UPDATED_CREDENTIALS
for the call to choose_xattr_handler(name)->getxattr. This caused
function SaunaClient::getxattr to fail on master restart.

Set sfsioretries=100 preventing the test to fail due to
strict test conditions(chunkservers and master restarting over time)

Signed-off-by: GigaCronos <jorge.cabrera@leil.io>