test(win): Adapt case insensitive test for Windows (#692)
Changes introduced on commit a0136ade included testing for hardlinks
and symlinks in a case insensitive manner. However, on Windows,
hardlinks and symlinks from Linux-style paths do not behave the same way
as on Linux systems, causing the tests to fail. This commit modifies the
test to skip hardlink and symlink tests when running on Windows systems.

Signed-off-by: rolysr <rolysr@leil.io>