chore(mount): Add string support for tweaks inode
This commit introduces code for supporting string type variables to be
register on a Tweaks class instance. The goal of this change is to add
support for some client string typed options to be updated on runtime
without the need of stopping and starting the client manually, instead
allow dynamic runtime string options update to take effect on client.

Signed-off-by: Rolando Sánchez Ramos <rolysr@leil.io>