ci: fix bzip2 from vcpkg
The port definition for bzip2 was resolved by versioning, which was
using the git protocol to download the resources. bzip2 is only used by
boost-iostreams, which is also only used when the cmake option
ENABLE_REQUEST_LOG is enabled (almost never).

The versioning way, was using git over https and was blocked
temporarily by sourceware due to DDoS attack, this change switches to a
not versioned port, allowing to download the needed resources from a
tar file.

Signed-off-by: guillex <guillex@leil.io>