chore: Add code signing to Windows compiled bins
As code signing procedure is added for Windows client installer, it
should also be applied to the Windows client binaries to ensure
consistency and  security across all distributed executables. This
change adds the signing script from Windows client specific code to the
.gitingnore and changes the way the Window binaries metadata is placed
during the build process to ensure this procedure is done before signing
the binaries.

Signed-off-by: rolysr <rolysr@leil.io>